% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigRadioButton.R
\name{bigRadioButtons}
\alias{bigRadioButtons}
\title{Radio buttons that display HTML}
\usage{
bigRadioButtons(id, label, choices, origNum)
}
\arguments{
\item{choices}{a character vector, one for each choice}
}
\description{
Radio buttons that display HTML
}
