% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_exercise.R
\name{new_calcZ_exercise}
\alias{new_calcZ_exercise}
\title{RStudio addin to make a new CalcZ book exercise}
\usage{
new_calcZ_exercise(directory = ".", exercise_id = NULL, save = TRUE)
}
\description{
RStudio addin to make a new CalcZ book exercise
}
