% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQfile.R
\docType{data}
\name{test1}
\alias{test1}
\title{test lines for development}
\format{
An object of class \code{character} of length 4.
}
\usage{
test1
}
\description{
test lines for development
}
\keyword{datasets}
